FEAT: 샵 인증 보호 라우트를 pathless 레이아웃으로 리팩토링
TanStack Router의 pathless 레이아웃(_auth)을 도입하여
인증이 필요한 라우트를 통합 관리합니다.

- 새 파일: routes/_auth.tsx - 인증 보호 레이아웃 라우트
  - 미인증 시 LoginBottomSheet 자동 표시
  - URL 경로 변경 없음 (_auth는 pathless)

- 6개 라우트를 routes/_auth/ 디렉토리로 이동:
  - my-orders.tsx
  - order.$orderNumber.tsx
  - cancel-detail.$orderNumber.tsx
  - cancel-request.$orderNumber.tsx
  - new-order.$orderNumber.tsx
  - order-complete.$orderNumber.tsx

- 각 라우트의 createFileRoute 경로 업데이트
  - 예: '/my-orders' → '/_auth/my-orders'

- 개별 라우트의 중복된 인증 체크 로직 제거
  - my-orders.tsx의 LoginRequiredView 제거
  - new-order.tsx의 AuthPromptSection 제거

이제 모든 인증 로직이 _auth 레이아웃에서 중앙 집중 관리됩니다.
